+++ assets_audio_player_web/lib/web/web_player_howl.dart
@@ -95,12 +95,12 @@ class WebPlayerHowl extends WebPlayer {
   }
 
   @override
-  Future<void> open({String path,
-    String audioType,
-    double volume,
-    double seek,
-    bool autoStart
-  }) async {
+  Future<void> open(
+      {String path,
+      String audioType,
+      double volume,
+      double seek,
+      bool autoStart}) async {
     stop();
 
     _howl = Howl(src: [findAssetPath(path, audioType)]);
@@ -109,7 +109,7 @@ class WebPlayerHowl extends WebPlayer {
       play();
     }
     this.volume = volume;
-    if(seek != null){
+    if (seek != null) {
       this.seek(to: seek);
     }
   }